requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile ../requirements.in
#
asgiref==3.8.1
    # via django
astroid==3.3.9
    # via pylint
async-timeout==5.0.1
    # via redis
attrs==25.3.0
    # via glom
black==25.1.0
    # via -r ../requirements.in
boltons==25.0.0
    # via
    #   face
    #   glom
certifi==2025.1.31
    # via requests
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via black
dill==0.3.9
    # via pylint
django==3.2.25
    # via
    #   -r ../requirements.in
    #   django-redis
django-redis==5.4.0
    # via -r ../requirements.in
exceptiongroup==1.2.2
    # via pytest
face==24.0.0
    # via glom
glom==24.11.0
    # via -r ../requirements.in
gunicorn==23.0.0
    # via -r ../requirements.in
idna==3.10
    # via requests
iniconfig==2.1.0
    # via pytest
isort==6.0.1
    # via
    #   -r ../requirements.in
    #   pylint
mccabe==0.7.0
    # via pylint
mypy-extensions==1.0.0
    # via black
mysqlclient==2.2.7
    # via -r ../requirements.in
packaging==24.2
    # via
    #   black
    #   gunicorn
    #   pytest
pathspec==0.12.1
    # via black
platformdirs==4.3.7
    # via
    #   black
    #   pylint
pluggy==1.5.0
    # via pytest
pylint==3.3.6
    # via -r ../requirements.in
pysolr==3.10.0
    # via -r ../requirements.in
pytest==8.3.5
    # via pytest-django
pytest-django==4.11.0
    # via -r ../requirements.in
python-dotenv==1.0.1
    # via -r ../requirements.in
pytz==2025.1
    # via django
redis==5.2.1
    # via
    #   -r ../requirements.in
    #   django-redis
requests==2.32.3
    # via
    #   -r ../requirements.in
    #   pysolr
sqlparse==0.5.3
    # via django
tomli==2.2.1
    # via
    #   black
    #   pylint
    #   pytest
tomlkit==0.13.2
    # via pylint
typing-extensions==4.12.2
    # via
    #   asgiref
    #   astroid
    #   black
urllib3==2.3.0
    # via requests

# The following packages are considered to be unsafe in a requirements file:
# setuptools
